backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==2.0.0
asyncpg==0.29.0
beautifulsoup4==4.12.3
blivet==3.8.2
blivet-gui==2.5.0
cairocffi==1.5.1
CairoSVG==2.7.0
catfish==4.18.0
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cssselect2==0.7.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
distro==1.8.0
dnf==4.19.2
dnspython==2.6.1
fastapi==0.110.0
file-magic==0.4.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
humanize==3.13.1
idna==3.7
image-go-nord==0.1.4
initial-setup==0.3.98
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.1
lxml==4.9.3
Mako==1.3.3
MarkupSafe==2.1.5
motor==3.3.2
nftables==0.1
olefile==0.46
packaging==23.1
passlib==1.7.4
pexpect==4.8.0
pid==2.2.3
pillow==10.3.0
ply==3.11
productmd==1.38
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
pycurl==7.45.2
pydantic==2.6.4
pydantic_core==2.16.3
pyenchant==3.2.2
PyGObject==3.46.0
PyJWT==2.8.0
pykickstart==3.48
pymongo==4.6.2
pyparted==3.13.0
PySocks==1.7.1
pystray==0.17.3
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-gettext==4.0
python-manatools==0.0.4
python-meh==0.51
python-xlib==0.33
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2024.4.16
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.19.1.1
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.33/src
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.6.0
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
systemd-python==235
tinycss2==1.2.1
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
